backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
pydantic==1.10.7
httpx==0.24.1
google-auth==2.17.3
httpx-oauth==0.10.0
twilio==8.0.0
python-dateutil==2.8.2
